Always update the services DB timestamp.
Had the problem that if the status of a service didn't change a lot then the update would not actually do anything and therefore the timestamp would always be old. This caused the web front end to report services as stuck. Also in this PR i change the logic to reduce by one the number of calls to the DB.

	modified:   scripts/ganga-daemon.py