Reload the table after admin change.
After the admin management fancybox closes we need to refresh the table.
This is because if a user demotes himself then he needs to nolonger have
the admin view. This requires a full reload rather than an ajax reload since
the table structure as well as context menus will be different as well as
it's contents.

	modified:   src/html/admins.html
	modified:   src/html/index.html