Refactoring the ganga-daemon.py file.
This commit moves the importing of ganga to inside the daemon_main.
This is because ganga need some file descriptors which would be closed
if imported before spawning the daemon process. I also take the chance to
refactor the ganga-daemon.py file to hopefully make it more readable and
maintainable.

	modified:   scripts/ganga-daemon.py