Fix security bug in RequestDBAPI PUT.
This bug meant that a user could change any DB column they liked, including the status to anything. This did not allow them to affect jobs other than their own but it did mean that they could screw up the system by changing fields that were never intended for the user to modify.

	modified:   src/python/frontend/RequestsDBAPI.py